# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFT Service",
    "description": "The service integrates with the NFT-token and allows you to interact with it",
    "termsOfService": "http://someurl.oxox",
    "contact": {
      "name": "Telegram: @pavel_gr21",
      "email": "vhser@yandex.ru"
    },
    "license": {
      "name": "MIT",
      "url": "https://en.wikipedia.org/wiki/MIT_License"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8008",
      "description": " Server-1"
    }
  ],
  "paths": {
...
    "/api/transfers/create": {
      "post": {
        "responses": {
          "201": {
            "description": "Successfully created transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Transfers"
        ],
        "description": " Creates a new transfer of the NFT token to a new owner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/api/transfers/list": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response containing the list of transfers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Transfers"
        ],
        "description": " Returns a list of transfers. If `limit` and `offset` parameters are provided, they will be used for pagination. By default, `limit` is set to 200, and `offset` is 0. The `limit` value must be between 1 and 500.",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset, default 0",
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "Pagination offset, default 0"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of pagination elements, default 200, max 500",
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "Number of pagination elements, default 200, max 500"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTransferRequest": {
        "type": "object",
        "properties": {
          "from_address": {
            "type": "string"
          },
          "to_address": {
            "type": "string"
          },
          "token_id": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "Transfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "from_address": {
            "type": "string"
          },
          "to_address": {
            "type": "string"
          },
          "token_id": {
            "type": "string"
          },
          "tx_hash": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "integer"
          },
          "updated_at": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}